+++ DawnLib/src/API/Terminal/TerminalKeywordBuilder.cs
@@ -2,41 +2,45 @@ using System;
 using System.Collections.Generic;
 using UnityEngine;
 using Dawn.Internal;
+using Dawn.Utils;
 
 namespace Dawn;
 public class TerminalKeywordBuilder
 {
+    internal static List<TerminalKeyword> _terminalKeywords = new List<TerminalKeyword>();
     private TerminalKeyword _keyword;
 
-    internal TerminalKeywordBuilder(string name, string word, ITerminalKeyword.DawnKeywordType _priority)
+    internal TerminalKeywordBuilder(string name, string word, ITerminalKeyword.DawnKeywordType keywordPriority)
     {
-        if (TerminalRefs.Instance.TryGetKeyword(word, out _keyword))
+        bool terminalExists = TerminalRefs.Instance != null;
+        if ((terminalExists && TerminalRefs.Instance!.TryGetKeyword(word, out TerminalKeyword existingKeywordWithSameWord)) || (!terminalExists && NoTerminalTryGetKeyword(word, out existingKeywordWithSameWord)))
         {
-            ITerminalKeyword.DawnKeywordType existingPriority = _keyword.GetKeywordPriority();
-            if (existingPriority <= _priority)
+            ITerminalKeyword.DawnKeywordType existingPriority = existingKeywordWithSameWord.GetKeywordPriority();
+            if (existingPriority <= keywordPriority)
             {
                 DawnPlugin.Logger.LogWarning($"'{word}' already has an existing TerminalKeyword with a higher priority [ {existingPriority} ]");
                 //below still creates a new keyword with just a unique (and unusuable) replacement word.
                 //throwing an exception here breaks the terminal and returning a null keyword will just break the terminal later
                 _keyword = ScriptableObject.CreateInstance<TerminalKeyword>();
                 _keyword.name = name;
-                _keyword.SetKeywordPriority(_priority);
+                _keyword.SetKeywordPriority(keywordPriority);
                 OverrideWord(word + _keyword.GetHashCode());
                 DawnPlugin.Logger.LogWarning($"TerminalKeyword word set to {_keyword.word}");
                 return;
             }
             else
             {
-                DawnPlugin.Logger.LogWarning($"Replacing keyword [{_keyword.word}] with priority [ {existingPriority} ] due to new keyword with the same word at a higher priority [ {_priority} ]");
+                _keyword = existingKeywordWithSameWord;
+                DawnPlugin.Logger.LogWarning($"Replacing keyword [{_keyword.word}] with priority [ {existingPriority} ] due to new keyword with the same word at a higher priority [ {keywordPriority} ]");
                 _keyword.name = name;
-                _keyword.SetKeywordPriority(_priority);
+                _keyword.SetKeywordPriority(keywordPriority);
                 return;
             }
         }
 
         _keyword = ScriptableObject.CreateInstance<TerminalKeyword>();
         _keyword.name = name;
-        _keyword.SetKeywordPriority(_priority);
+        _keyword.SetKeywordPriority(keywordPriority);
         OverrideWord(word);
     }
 
@@ -72,7 +76,7 @@ public class TerminalKeywordBuilder
         return this;
     }
 
-    public TerminalKeywordBuilder AccessTerminalObjects(bool accessTerminalObjects)
+    public TerminalKeywordBuilder SetAccessTerminalObjects(bool accessTerminalObjects)
     {
         _keyword.accessTerminalObjects = accessTerminalObjects;
         return this;
@@ -100,5 +104,21 @@ public class TerminalKeywordBuilder
     public TerminalKeyword Build()
     {
         return _keyword;
+    }
+
+    public static bool NoTerminalTryGetKeyword(string keyWord, out TerminalKeyword terminalKeyword)
+    {
+        foreach (TerminalKeyword keyword in _terminalKeywords)
+        {
+            if (keyWord.CompareStringsInvariant(keyword.word))
+            {
+                //Loggers.LogDebug($"Keyword: [{keyWord}] found!");
+                terminalKeyword = keyword;
+                return true;
+            }
+        }
+
+        terminalKeyword = null!;
+        return false;
     }
 }